+++ authorization/service/helpers.go
@@ -1,0 +1,63 @@
+package main
+
+import (
+	"context"
+	"database/sql"
+	"fmt"
+	"golang.org/x/crypto/bcrypt"
+	"log"
+	pb "miracle/proto"
+)
+
+func (s *server) ValidateEmail(ctx context.Context, req *pb.EmailValidationRequest) (*pb.EmailValidationResponse, error) {
+	query := "SELECT COUNT(*) FROM users WHERE email = $1"
+	var count int
+	err := s.db.QueryRow(query, req.Email).Scan(&count)
+	if err != nil {
+		return nil, fmt.Errorf("failed to validate email: %v", err)
+	}
+
+	response := &pb.EmailValidationResponse{
+		Valid: count == 0,
+	}
+
+	return response, nil
+}
+
+func generateUniqueIDforUser(db *sql.DB) (int32, error) {
+	var id int32
+	err := db.QueryRow("SELECT COALESCE(MAX(id), 0) + 1 FROM users").Scan(&id)
+	if err != nil {
+		log.Printf("Failed to generate unique ID: %v", err)
+		return 0, err
+	}
+	return id, nil
+}
+
+func generateUniqueIDforCar(db *sql.DB) (int32, error) {
+	var id int32
+	err := db.QueryRow("SELECT COALESCE(MAX(id), 0) + 1 FROM car").Scan(&id)
+	if err != nil {
+		log.Printf("Failed to generate unique ID: %v", err)
+		return 0, err
+	}
+	return id, nil
+}
+
+func generateUniqueIDforRent(db *sql.DB) (int32, error) {
+	var id int32
+	err := db.QueryRow("SELECT COALESCE(MAX(id), 0) + 1 FROM rented_cars").Scan(&id)
+	if err != nil {
+		log.Printf("Failed to generate unique ID: %v", err)
+		return 0, err
+	}
+	return id, nil
+}
+
+func hashPassword(password string) (string, error) {
+	hash, err := bcrypt.GenerateFromPassword([]byte(password), bcrypt.DefaultCost)
+	if err != nil {
+		return "", err
+	}
+	return string(hash), nil
+}
